+++ Changelog.md
@@ -34,6 +34,9 @@
 1. Access the original parsed version of an SDF document with `Element::OriginalVersion`.
     * [Pull request 640](https://bitbucket.org/osrf/sdformat/pull-requests/640)
 
+1. Model::Load: fail fast if an sdf 1.7 file has name collisions.
+    * [Pull request 648](https://bitbucket.org/osrf/sdformat/pull-requests/648)
+
 ### SDFormat 9.0.0 (2019-12-10)
 
 1. Move recursiveSameTypeUniqueNames from ign.cc to parser.cc and make public.